+++ src/main/java/org/folio/rest/impl/InitAPIs.java
@@ -1,5 +1,14 @@
 package org.folio.rest.impl;
 
+import io.vertx.core.AsyncResult;
+import io.vertx.core.Context;
+import io.vertx.core.Future;
+import io.vertx.core.Handler;
+import io.vertx.core.Vertx;
+import io.vertx.serviceproxy.ServiceBinder;
+import org.folio.rest.resource.interfaces.InitAPI;
+import org.folio.services.PasswordStorageService;
+
 import io.vertx.core.*;
 import io.vertx.serviceproxy.ServiceBinder;
 import org.folio.rest.resource.interfaces.InitAPI;
@@ -8,6 +17,8 @@ import org.folio.services.LogStorageService;
 
 import java.net.URL;
 import java.util.MissingResourceException;
+
+import static org.folio.util.LoginConfigUtils.PW_CONFIG_PROXY_STORY_ADDRESS;
 
 import static org.folio.util.LoginConfigUtils.EVENT_CONFIG_PROXY_CONFIG_ADDRESS;
 import static org.folio.util.LoginConfigUtils.EVENT_CONFIG_PROXY_STORY_ADDRESS;
@@ -23,9 +34,12 @@ public class InitAPIs implements InitAPI {
 
   public void init(Vertx vertx, Context context, Handler<AsyncResult<Boolean>> resultHandler) {
     URL u = InitAPIs.class.getClassLoader().getResource(CREDENTIAL_SCHEMA_PATH);
-    if(u == null) {
+    if (u == null) {
       resultHandler.handle(Future.failedFuture(new MissingResourceException(CREDENTIAL_SCHEMA_PATH, InitAPIs.class.getName(), CREDENTIAL_SCHEMA_PATH)));
     } else {
+      new ServiceBinder(vertx)
+        .setAddress(PW_CONFIG_PROXY_STORY_ADDRESS)
+        .register(PasswordStorageService.class, PasswordStorageService.create(vertx));
       new ServiceBinder(vertx)
         .setAddress(EVENT_CONFIG_PROXY_STORY_ADDRESS)
         .register(LogStorageService.class, LogStorageService.create(vertx));
@@ -36,4 +50,5 @@ public class InitAPIs implements InitAPI {
       resultHandler.handle(Future.succeededFuture(true));
     }
   }
+
 }