+++ src/main/java/org/folio/util/LoginConfigUtils.java
@@ -14,10 +14,17 @@ public class LoginConfigUtils {
    */
   public static final String EVENT_CONFIG_PROXY_STORY_ADDRESS = "event-store-service.queue";
   public static final String EVENT_CONFIG_PROXY_CONFIG_ADDRESS = "event-config-service.queue";
+  public static final String PW_CONFIG_PROXY_STORY_ADDRESS = "password-store-service.queue";
 
   public static final String EVENT_LOG_API_MODULE = "EVENT_LOG";
   public static final String EVENT_LOG_API_CODE_STATUS = "STATUS";
   public static final String SNAPSHOTS_TABLE_EVENT_LOGS = "event_logs";
+
+  /**
+   * Tables
+   */
+  public static final String SNAPSHOTS_TABLE_PW = "auth_password_action";
+  public static final String SNAPSHOTS_TABLE_CREDENTIALS = "auth_credentials";
 
   public static final String VALUE_IS_NOT_FOUND = "isNotFound";
   public static final JsonObject EMPTY_JSON_OBJECT = new JsonObject().put(VALUE_IS_NOT_FOUND, true);